# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ScyllaDB Vector Store API",
    "description": "REST API for ScyllaDB Vector Store nodes. Provides capabilities for executing vector search queries, managing indexes, and checking status of Vector Store nodes.",
    "license": {
      "name": "LicenseRef-ScyllaDB-Source-Available-1.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/indexes": {
      "get": {
        "tags": [
          "scylla-vector-store-index"
        ],
        "description": "Returns the list of indexes managed by the Vector Store node. The list includes indexes in any state (initializing, available/built, destroying). Due to synchronization delays, it may temporarily differ from the list of vector indexes inside ScyllaDB.",
        "operationId": "get_indexes",
        "responses": {
          "200": {
            "description": "Successful operation. Returns an array of index information representing all indexes managed by the Vector Store.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IndexInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/indexes/{keyspace}/{index}/ann": {
      "post": {
        "tags": [
          "scylla-vector-store-index"
        ],
        "description": "Performs an Approximate Nearest Neighbor (ANN) search using the specified index. Returns the vectors most similar to the provided vector. The maximum number of results is controlled by the optional 'limit' parameter in the payload. The similarity metric is determined at index creation and cannot be changed per query.",
        "operationId": "post_index_ann",
        "parameters": [
          {
            "name": "keyspace",
            "in": "path",
            "description": "The name of the ScyllaDB keyspace containing the vector index.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/KeyspaceName"
            }
          },
          {
            "name": "index",
            "in": "path",
            "description": "The name of the ScyllaDB vector index within the specified keyspace to perform the search on.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IndexName"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostIndexAnnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful ANN search. Returns a list of primary keys and their corresponding distances for the most similar vectors found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostIndexAnnResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Possible causes: invalid vector size, malformed input, or missing required fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Index not found. Possible causes: index does not exist, or is not discovered yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Error while searching vectors. Possible causes: internal error, or search engine issues.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable. Indicates that a full scan of the index is in progress and the search cannot be performed at this time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/indexes/{keyspace}/{index}/count": {
      "get": {
        "tags": [
          "scylla-vector-store-index"
        ],
        "description": "Returns the number of vectors indexed by a specific vector index. Reflects only available vectors and excludes any 'tombstones' (elements marked for deletion but still present in the index structure).",
        "operationId": "get_index_count",
        "parameters": [
          {
            "name": "keyspace",
            "in": "path",
            "description": "The name of the ScyllaDB keyspace containing the vector index.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/KeyspaceName"
            }
          },
          {
            "name": "index",
            "in": "path",
            "description": "The name of the ScyllaDB vector index within the specified keyspace to count vectors for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IndexName"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful count operation. Returns the total number of vectors currently stored in the index.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            }
          },
          "404": {
            "description": "Index not found. Possible causes: index does not exist, or is not discovered yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "Error while counting vectors. Possible causes: internal error, or issues accessing the database.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/info": {
      "get": {
        "tags": [
          "scylla-vector-store-info"
        ],
        "description": "Returns information about the Vector Store service serving this API.",
        "operationId": "get_info",
        "responses": {
          "200": {
            "description": "Vector Store service information.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/status": {
      "get": {
        "tags": [
          "scylla-vector-store-info"
        ],
        "description": "Returns the current operational status of the Vector Store node.",
        "operationId": "get_status",
        "responses": {
          "200": {
            "description": "Successful operation. Returns the current operational status of the Vector Store node.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DataType": {
        "type": "string",
        "description": "Data type and precision used for storing and processing vectors in the index.",
        "enum": [
          "F32"
        ]
      },
      "Distance": {
        "type": "number",
        "format": "float",
        "description": "Distance between vectors measured using the distance function defined while creating the index."
      },
      "ErrorMessage": {
        "type": "string",
        "description": "A human-readable description of the error that occurred."
      },
      "IndexInfo": {
        "type": "object",
        "description": "Information about a vector index, such as keyspace, name and data type.",
        "required": [
          "keyspace",
          "index",
          "data_type"
        ],
        "properties": {
          "data_type": {
            "$ref": "#/components/schemas/DataType"
          },
          "index": {
            "$ref": "#/components/schemas/IndexName"
          },
          "keyspace": {
            "$ref": "#/components/schemas/KeyspaceName"
          }
        }
      },
      "IndexName": {
        "type": "string",
        "description": "A name of the vector index in a db."
      },
      "InfoResponse": {
        "type": "object",
        "required": [
          "version",
          "service"
        ],
        "properties": {
          "service": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "KeyspaceName": {
        "type": "string",
        "description": "A keyspace name in a db."
      },
      "Limit": {
        "type": "integer",
        "format": "int32"
      },
      "PostIndexAnnRequest": {
        "type": "object",
        "required": [
          "vector"
        ],
        "properties": {
          "limit": {
            "$ref": "#/components/schemas/Limit"
          },
          "vector": {
            "$ref": "#/components/schemas/Vector"
          }
        }
      },
      "PostIndexAnnResponse": {
        "type": "object",
        "required": [
          "primary_keys",
          "distances"
        ],
        "properties": {
          "distances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Distance"
            }
          },
          "primary_keys": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {}
            },
            "propertyNames": {
              "type": "string",
              "description": "Name of the column in a db table."
            }
          }
        }
      },
      "Status": {
        "type": "string",
        "description": "Operational status of the Vector Store node.",
        "enum": [
          "INITIALIZING",
          "CONNECTING_TO_DB",
          "BOOTSTRAPPING",
          "SERVING"
        ],
        "x-enum-descriptions": [
          "The node is starting up.",
          "The node is establishing a connection to ScyllaDB.",
          "The node is discovering available vector indexes in ScyllaDB.",
          "The node is making initial indexes discovering and vector data indexing."
        ]
      },
      "Vector": {
        "type": "array",
        "items": {
          "type": "number",
          "format": "float"
        },
        "description": "The vector to use for the Approximate Nearest Neighbor search. The format of data must match the data_type of the index."
      }
    }
  },
  "tags": [
    {
      "name": "scylla-vector-store-index",
      "description": "Operations for managing ScyllaDB Vector Store indexes, including listing, counting, and searching."
    },
    {
      "name": "scylla-vector-store-info",
      "description": "Endpoints providing general information and status about the ScyllaDB Vector Store service."
    }
  ]
}